server/pyproject.toml
[project]
name = "hearmypaper-server"
version = "0.1.0"
description = "HearMyPaper server"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.30.0",
    "cryptography>=45.0.7",
    "fastapi[standard]>=0.116.1",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.10.1",
    "pyjwt>=2.10.1",
    "sqlalchemy>=2.0.43",
]

[project.optional-dependencies]
dev = [
    "mypy>=1.18.2",
    "pre-commit>=4.3.0",
    "ruff>=0.13.1",
]
